Restore ending new line of README.rst
Since some days ago we've been receiving this warning when
running `bt test code-quality`:

`[W7908(missing-newline-extrafiles), ]  Missing newline`

It's because the template file for our readme is missing the
very last end of line, at the end of the file. It was removed
because of github.com/brain-tec/maintainer-tools/pull/67/files,
probably because of a missconfiguration of PyCharm, set up to
remove trailing blank lines at the end of each saved file.

The setting is in "File" > "Settings" > "Editor" > "General" >
"Remove trailing blank lines at the end of save files", and the
DEV making the aforementioned pull probably had it activated.